+++ src/components/input-field/examples/input-field-text.tsx
@@ -16,16 +16,16 @@ export class InputFieldTextExample {
     private disabled = false;
 
     @State()
+    private readonly = false;
+
+    @State()
     private invalid = false;
 
     @State()
-    private value;
+    private value: string;
 
     constructor() {
         this.checkValidity = this.checkValidity.bind(this);
-        this.changeHandler = this.changeHandler.bind(this);
-        this.toggleEnabled = this.toggleEnabled.bind(this);
-        this.toggleRequired = this.toggleRequired.bind(this);
     }
 
     public render() {
@@ -40,17 +40,25 @@ export class InputFieldTextExample {
                 required={this.required}
                 invalid={this.invalid}
                 disabled={this.disabled}
+                readonly={this.readonly}
                 onChange={this.changeHandler}
             />,
             <p>
                 <limel-flex-container justify="end">
-                    <limel-button
-                        onClick={this.toggleEnabled}
-                        label={this.disabled ? 'Enable' : 'Disable'}
+                    <limel-checkbox
+                        checked={this.disabled}
+                        label="Disabled"
+                        onChange={this.setDisabled}
                     />
-                    <limel-button
-                        onClick={this.toggleRequired}
-                        label={this.required ? 'Set Optional' : 'Set Required'}
+                    <limel-checkbox
+                        checked={this.readonly}
+                        label="Readonly"
+                        onChange={this.setReadonly}
+                    />
+                    <limel-checkbox
+                        checked={this.required}
+                        label="Required"
+                        onChange={this.setRequired}
                     />
                 </limel-flex-container>
             </p>,
@@ -63,16 +71,22 @@ export class InputFieldTextExample {
         this.invalid = this.required && !this.value;
     }
 
-    private changeHandler(event) {
+    private changeHandler = (event: CustomEvent<string>) => {
         this.value = event.detail;
-        this.checkValidity();
-    }
+    };
 
-    private toggleEnabled() {
-        this.disabled = !this.disabled;
-    }
+    private setDisabled = (event: CustomEvent<boolean>) => {
+        event.stopPropagation();
+        this.disabled = event.detail;
+    };
 
-    private toggleRequired() {
-        this.required = !this.required;
-    }
+    private setReadonly = (event: CustomEvent<boolean>) => {
+        event.stopPropagation();
+        this.readonly = event.detail;
+    };
+
+    private setRequired = (event: CustomEvent<boolean>) => {
+        event.stopPropagation();
+        this.required = event.detail;
+    };
 }