+++ src/components/input-field/examples/input-field-textarea.tsx
@@ -6,6 +6,7 @@ import { Component, h, State } from '@stencil/core';
 @Component({
     tag: 'limel-example-input-field-textarea',
     shadow: true,
+    styleUrl: 'input-field.scss',
 })
 export class InputFieldTextareaExample {
     @State()
@@ -15,16 +16,13 @@ export class InputFieldTextareaExample {
     private disabled = false;
 
     @State()
+    private readonly = false;
+
+    @State()
     private value: string;
 
-    constructor() {
-        this.changeHandler = this.changeHandler.bind(this);
-        this.setEnabled = this.setEnabled.bind(this);
-        this.setRequired = this.setRequired.bind(this);
-    }
-
     public render() {
-        const MAX_LENGTH = 240;
+        const MAX_LENGTH = 500;
 
         return [
             <limel-input-field
@@ -34,14 +32,19 @@ export class InputFieldTextareaExample {
                 maxlength={MAX_LENGTH}
                 value={this.value}
                 required={this.required}
-                disabled={this.disabled}
                 onChange={this.changeHandler}
+                disabled={this.disabled}
+                readonly={this.readonly}
             />,
             <p>
                 <limel-flex-container justify="end">
                     <limel-checkbox
-                        onChange={this.setEnabled}
+                        onChange={this.setDisabled}
                         label="Disabled"
+                    />
+                    <limel-checkbox
+                        onChange={this.setReadonly}
+                        label="Readonly"
                     />
                     <limel-checkbox
                         onChange={this.setRequired}
@@ -53,17 +56,22 @@ export class InputFieldTextareaExample {
         ];
     }
 
-    private changeHandler(event) {
+    private changeHandler = (event: CustomEvent<string>) => {
         this.value = event.detail;
-    }
+    };
 
-    private setEnabled(event: CustomEvent<boolean>) {
+    private setDisabled = (event: CustomEvent<boolean>) => {
         event.stopPropagation();
         this.disabled = event.detail;
-    }
+    };
 
-    private setRequired(event: CustomEvent<boolean>) {
+    private setReadonly = (event: CustomEvent<boolean>) => {
+        event.stopPropagation();
+        this.readonly = event.detail;
+    };
+
+    private setRequired = (event: CustomEvent<boolean>) => {
         event.stopPropagation();
         this.required = event.detail;
-    }
+    };
 }